+++ src/app/components/Hero/Guest/index.ts
@@ -1,0 +1,2 @@
+export { AuthForm } from "./AuthForm";
+export { JournalPic } from "./JournalPic";
