+++ src/app/components/Hero/LoggedInUser/QuoteWrapper.tsx
