+++ src/app/components/Hero/LoggedInUser/index.ts
@@ -1,0 +1,2 @@
+export { LemonadePic } from "./LemonadePic";
+export { Quote } from "./Quote";
